+++ view/frontend/web/js/view/payment/nuvei.js
@@ -20,9 +20,8 @@ define(
 			window.nuveiCheckoutSdk	= checkout;
 			window.checkout			= magentoTmpCheckout;
 		};
-		nuveiCheckoutSdkScr.src = 1 == window.checkoutConfig.payment['nuvei'].useDevSdk
-            ? 'https://srv-bsf-devpppjs.gw-4u.com/checkoutNext/checkout.js'
-                : 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
+		
+        nuveiCheckoutSdkScr.src = 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
                 
 		document.head.appendChild(nuveiCheckoutSdkScr);
 		// /Load Nuvei Chekout SDK and add it ot a local variable