+++ view/frontend/web/js/view/payment/nuvei.js
@@ -13,17 +13,17 @@ define(
         'use strict';
         
 		// Load Nuvei Chekout SDK and add it ot a local variable
-		var magentoTmpCheckout	= window.checkout;
-		var nuveiCheckoutSdkScr	= document.createElement('script');
-        
-		nuveiCheckoutSdkScr.onload = function () {
-			window.nuveiCheckoutSdk	= checkout;
-			window.checkout			= magentoTmpCheckout;
-		};
-		
-        nuveiCheckoutSdkScr.src = 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
-                
-		document.head.appendChild(nuveiCheckoutSdkScr);
+//		var magentoTmpCheckout	= window.checkout;
+//		var nuveiCheckoutSdkScr	= document.createElement('script');
+//        
+//		nuveiCheckoutSdkScr.onload = function () {
+//			window.nuveiCheckoutSdk	= checkout;
+//			window.checkout			= magentoTmpCheckout;
+//		};
+//		
+//        nuveiCheckoutSdkScr.src = 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
+//                
+//		document.head.appendChild(nuveiCheckoutSdkScr);
 		// /Load Nuvei Chekout SDK and add it ot a local variable
 		
         rendererList.push({